moved note about replacement patterns to the replacement pattern section
I had a terrible time trying to figure out how to avoid triggering the replacement pattern, but didn't read the replacement function section because I didn't know it was relevant to replacement patterns, and wasn't something I needed. Makes more sense to me that the note be moved to the section about replacement patterns so users reading about them will have all the information. Perhaps it could be in both, but it would be a bit redundant and they're close together, so I think it's better to put it above where it's slightly more relevant.